+++ packages/nitro_annotations/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.2.2
+
+- Standardized annotation class metadata for Nitrogen 0.2.2's stable code generation.
+
 # 0.2.1
 
 - Created `nitro_annotations` package to house all Nitro annotations.